thayhuy/code_python/requirements.txt
Flask==2.0.1
paho-mqtt==1.6.1
requests==2.26.0
python-dotenv==0.19.0